post_boot.sh: Round off the interactive governor timer period
The interactive governor uses jiffies based timer for sampling the load.
It throws the below warning when the timer period is set to a value that
is not aligned to jiffies. Fix this by rounding off the timer period.

<4>[   26.540874] timer_rate not aligned to jiffy. Rounded up to 30000

This change should not make any difference. Because the timer period
is rounded off any way in the kernel.

Change-Id: I6258954650776bd30459d1de157f5869a55886a1